Rewrite the build system.
This replaces the build system with one like the autoconf standard.
That includes a configure script with all the usual options for setting
the prefix, enabling debugging, etc. The script uses pkg-config to find
the appropriate flags for the platform (with no-pkg-config fallback).
The new Makefile likewise has the GNU-recommended targets. Installation
now uses `install` instead of plain copy and chmod. Source dependencies
are now automatically generated during compilation.