Fix -Wshadow warnings.
I usually use this flag, so it got copied together with the build system
files. In berry there were a few instances of variable shadowing,
notably the c variable in client.c containing the command table.
I renamed it to a more descriptive "command_table" and made it const.
There also were a couple of signed/unsigned comparison warnings.