Add optional ID parameter for toggle_decorations
This alters the functionality of the toggle_decorations ipc command,
allowing it to now take an optional ID parameter.
If the parameter is specific, it will attempt to toggle decorations
only for that window, otherwise it will attempt to toggle decorations for
the current window (which is the current behavior).