Fix memory leak in ewmh_set_desktop_names
Free members of char** list at the end of ewmh_set_desktop_names. list
contains workspace numbers in text form and only the list pointer was
freed at the end, leaking the members.

Use asprintf instead of malloc/sprintf to print the numbers. malloc was
used to allocate a fixed size 3 byte array, which could overflow if
WORKSPACE_NUMBER were to be increased to 100.

Use calloc to allocate list so that in case of asprintf failure the
members would be NULL instead of random memory contents.