Avoid creation of zombie process by load_config
Previously, a zombie process was created when the process forked in
load_config terminates because the parent does not call wait or waitpid.
It is unreasonable to wait for the child because a user may exec their
autostart script into a long-running process. Instead, the parent now
simply ignores SIGCHLD which allows the child to be reaped immediately
upon termination.